renamed status package to dialogs because it contains dialogs
