+++ app/src/main/java/com/babylon/wallet/android/presentation/settings/SettingsNavGraph.kt
@@ -6,6 +6,8 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
 import com.babylon.wallet.android.BuildConfig
+import com.babylon.wallet.android.presentation.dialogs.assets.fungibleAssetDialog
+import com.babylon.wallet.android.presentation.dialogs.assets.nftAssetDialog
 import com.babylon.wallet.android.presentation.navigation.Screen
 import com.babylon.wallet.android.presentation.settings.approveddapps.approvedDAppsScreen
 import com.babylon.wallet.android.presentation.settings.approveddapps.dappdetail.dAppDetailScreen
@@ -19,8 +21,6 @@ import com.babylon.wallet.android.presentation.settings.securitycenter.securityC
 import com.babylon.wallet.android.presentation.settings.securitycenter.securityCenterNavGraph
 import com.babylon.wallet.android.presentation.settings.troubleshooting.troubleshootingNavGraph
 import com.babylon.wallet.android.presentation.settings.troubleshooting.troubleshootingSettings
-import com.babylon.wallet.android.presentation.status.assets.fungibleAssetDialog
-import com.babylon.wallet.android.presentation.status.assets.nftAssetDialog
 
 @Suppress("LongMethod")
 fun NavGraphBuilder.settingsNavGraph(
